Replaced size_t with uint64_t for Bitfield

+++ source/base/memory/bitfield.inl
@@ -1,12 +1,13 @@
 #ifndef SU_BASE_MEMORY_BITFIELD_INL
 #define SU_BASE_MEMORY_BITFIELD_INL
 
-#include <cstring>
 #include "bitfield.hpp"
+
+#include <cstring>
 
 namespace memory {
 
-inline Bitfield::Bitfield(size_t num_bits) noexcept
+inline Bitfield::Bitfield(uint64_t num_bits) noexcept
     : num_bytes_(num_bytes(num_bits)),
       buffer_(new uint32_t[num_bytes(num_bits) / sizeof(uint32_t)]) {}
 
@@ -14,7 +15,7 @@ inline Bitfield::~Bitfield() noexcept {
     delete[] buffer_;
 }
 
-inline size_t Bitfield::num_bytes() const noexcept {
+inline uint64_t Bitfield::num_bytes() const noexcept {
     return num_bytes_;
 }
 
@@ -22,7 +23,7 @@ inline void Bitfield::clear() noexcept {
     std::memset(buffer_, 0, num_bytes_);
 }
 
-inline void Bitfield::set(size_t index, bool value) noexcept {
+inline void Bitfield::set(uint64_t index, bool value) noexcept {
     uint32_t const mask = Mask >> (index % Bits);
 
     if (value) {
@@ -32,7 +33,7 @@ inline void Bitfield::set(size_t index, bool value) noexcept {
     }
 }
 
-inline bool Bitfield::get(size_t index) const noexcept {
+inline bool Bitfield::get(uint64_t index) const noexcept {
     uint32_t const mask = Mask >> (index % Bits);
 
     return (buffer_[index >> Log2Bits] & mask) != 0;
@@ -42,8 +43,8 @@ inline uint32_t* Bitfield::data() const noexcept {
     return buffer_;
 }
 
-inline size_t Bitfield::num_bytes(size_t num_bits) noexcept {
-    size_t const chunks = num_bits / Bits;
+inline uint64_t Bitfield::num_bytes(uint64_t num_bits) noexcept {
+    uint64_t const chunks = num_bits / Bits;
     return ((num_bits % Bits == 0) ? chunks : chunks + 1) * sizeof(uint32_t);
 }
